fix: split linux input subsystem constants
