+++ Cargo.toml
@@ -25,7 +25,7 @@ members = [
     "modules/rendering-configuration",
     "modules/asset-processing",
     "modules/native-io",
-    "base-lib/pinned-futures-helper", "platform/linux/epoll",
+    "base-lib/pinned-futures-helper",
 ]
 
 [workspace.dependencies]
