+++ platform/linux/input/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "linux-input"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+libc = "0.2"
